pyd7a-master/requirements.txt
bitstring
Cerberus==1.1
#eventlet
#Flask
#Flask-SocketIO
#Flask-Cors
PyCRC
pyserial>=3.4.0
unittest2
coverage
enum34
jsonpickle
paho-mqtt
traceback2
cogapp
